feat: getting user's task, update task status, delete task API 추가

+++ src/tasks/tasks.service.ts
@@ -4,14 +4,14 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-tasks-filter.dto';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
-import { TaskReository } from './task.repository';
+import { TaskRepository } from './task.repository';
 import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
   constructor(
-    @InjectRepository(TaskReository)
-    private taskRepository: TaskReository,
+    @InjectRepository(TaskRepository)
+    private taskRepository: TaskRepository,
   ) {}
 
   getTasks(
@@ -20,9 +20,12 @@ export class TasksService {
     return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
-    if (!found) {
+  async getTaskById(
+    id: number,
+    user: User
+    ): Promise<Task> {
+      const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
+      if (!found) {
       throw new NotFoundException(`Task with ID '${id}' not found.`);
     }
     return found;
@@ -31,16 +34,18 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(
+    id: number,
+    user: User): Promise<void> {
+      const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID '${id}' not found.`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus) {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user : User) {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
